Fix circular import in classification
